requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Brotli==1.0.9
cchardet==2.1.7
cffi==1.15.0
chardet==4.0.0
ciso8601==2.2.0
colorlog==6.5.0
ed25519==1.5
hikari @ git+https://github.com/hikari-py/hikari@1acaee734e98378b35d5223e624b6b8383f96ee3
hikari-lightbulb==1.5.2
idna==3.3
multidict==5.2.0
pure25519==0.0.1
pycares==4.1.2
pycparser==2.20
typing-extensions==3.10.0.2
uvloop==0.16.0
yarl==1.7.0
